Untested (WIP) - Live Shader Test
In this experiment I'm going to try and create an opengl program that allows the user to modify shader code on the fly.

This will be exposed through a simple ImGUI text box that the user can enter the shader code into for both the vertex and fragment shader.

Can be extended in future, but for now it assumes that the uniforms aren't going to change much.

The idea is that the user types, then after some configured delay (measured in gametime elapsed) the shader code recompiles.