Add Functionality for resizing framebuffer
Currently when we were resizing the window, we were only resizing the viewport. However, because we are rendering through a framebuffer, the texture for the color attachment, and the depth renderbuffer did not have their size adjusted.

As a result, we would see strange artifacts when resizing the window which isn't desirable. Let's fix that by making sure that whenever we call back to the resize of the window, we resize any textures attached to our rendering frame buffer.