Add render feature to auto shader
Previously our autoshader couldn't render anything, now we add that functionality into the renderer. We also add some test cases, with the ability to "harness" open gl for testing.

Ideally, we could hook into the API to mock out the actual creation of windows here but that is probably more effort than it's worth.