Add Scissor Test around Auto Shader Box
This is a simple addition to the rendering that means that we only render the viewport around the auto shader box, and not underneath it as well.

This may or may not be the desired effect. We could potentially put something else here in it's place, but I think this is what I originally had in mind.