Use shader program through autoshader
As a first step to having our autoshader used, well, automatically we should use the auto shader as our primary shader program when rendering.

While we're at it, I realise I've been lazy by just using unique_ptrs everywhere to get around the C++ move semantics and treating everything like a reference ala java. Let's improve this. To start with, we've made the Shader owned by the AutoShader itself. Importantly, this was a lesson in the move assignment overload. Because we want to create in place without copying when compiling a new shader, it's important that we don't copy construct the autoshader.

OpenGL has it's draw backs in having a full state machine, but it's helping us learn how to be careful with unnecessary copying. IN this case, copy constructing a Shader object is a mistake because once the copy goes out of scope, the destructor on the copy is called. This destructor deletes the shader program. Since the shader program is a singleton, this means our reference to the Shader program has nothing in it.  Using the move assignment means that we can safely swap the important contents of the Shader (the handle), and avoid destroying our sacred program.